Fix de plusieurs bug

See merge request somon/discoodle!38

+++ api/src/main/java/com/discoodle/api/request/EditGroupRequest.java
@@ -10,7 +10,7 @@ import lombok.ToString;
 @EqualsAndHashCode
 @ToString
 public class EditGroupRequest {
-    private final Integer groups_id;
+    private final Long groups_id;
     private final String name;
     private final String description;
 }
